refactor: rename message.tsx to MessageBubble.tsx and consolidate message-related components for better organization
